[widgets] Move HostPort into separate file under widgets
